requirements.txt
py_clob_client
py_order_utils

supabase
gotrue

openai
httpx

eth-utils
eth-account
eth-keys

matplotlib

aiohttp
requests
python-dotenv
cryptography

sentence-transformers
pinecone
instructor

redis